feat: handle multiple authors; refactor: move parserFunc to seperate file
